Reshuffle exports into own directory
